requirements.txt
# ===================================
# Core Framework
# ===================================
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20

# ===================================
# HTTP Client (para API externa)
# ===================================
httpx==0.28.1

# ===================================
# Machine Learning - Core
# ===================================
numpy==1.26.4
opencv-python==4.10.0.84
scikit-learn==1.5.2
joblib==1.4.2

# ===================================
# Deep Learning - PyTorch + Transformers
# ===================================
torch==2.5.1
torchvision==0.20.1
transformers==4.46.3

# ===================================
# Deep Learning - TensorFlow/Keras
# ===================================
tensorflow==2.18.0
keras==3.7.0

# ===================================
# Computer Vision - Detection
# ===================================
mediapipe==0.10.20
face-recognition==1.3.0
dlib==19.24.6
ultralytics==8.3.51  # YOLO World

# ===================================
# Image Processing
# ===================================
Pillow==11.0.0
albumentations==1.4.23

# ===================================
# OpenAI (para /ask endpoint)
# ===================================
openai==1.57.2

# ===================================
# CORS
# ===================================
python-cors==1.0.0

# ===================================
# Utilities
# ===================================
python-dotenv==1.0.1

# ===================================
# Optional - Para desarrollo
# ===================================
# pytest==8.3.4
# pytest-asyncio==0.24.0
# black==24.10.0
# flake8==7.1.1
# ipython==8.30.0

# ===================================
# Notas de Instalación
# ===================================
# 
# Para instalar:
# pip install -r requirements.txt
#
# Si tienes GPU NVIDIA y quieres usar CUDA:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# Si tienes problemas con dlib en Windows:
# pip install dlib-binary
#
# Para Mac M1/M2:
# pip install tensorflow-macos
# pip install tensorflow-metal